rustc_resolve: move export recording into its own module.
